refactor: constants 나라이름 분리
